+++ src/picasapy/app/i18n/picasapy_hu.ts
@@ -4572,6 +4572,34 @@ Az arcfelismerés ettől független: egy mappa lehet figyelt úgy is, hogy az ar
         <source>Choose destination folder...</source>
         <translation>Célmappa kiválasztása…</translation>
     </message>
+    <message>
+        <source>%1 / %2 imported at %3/sec</source>
+        <translation>%1 / %2 importálva, %3/mp</translation>
+    </message>
+    <message>
+        <source>bytes</source>
+        <translation>bájt</translation>
+    </message>
+    <message>
+        <source>File types:</source>
+        <translation>Fájltípusok:</translation>
+    </message>
+    <message>
+        <source>Picture and Movie Files</source>
+        <translation>Kép- és filmfájlok</translation>
+    </message>
+    <message>
+        <source>Picture Files</source>
+        <translation>Képfájlok</translation>
+    </message>
+    <message>
+        <source>All Files</source>
+        <translation>Minden fájl</translation>
+    </message>
+    <message>
+        <source>Recent sources</source>
+        <translation>Korábbi források</translation>
+    </message>
 </context>
 <context>
     <name>TimelineController</name>